little improvement

+++ ting-global-vue-site/src/store/user.js
@@ -17,6 +17,7 @@ export default {
     state: {
         user: null,
         token: null,
+        templates: {},
         selectedTemplate: null
     },
     mutations: {
@@ -31,7 +32,10 @@ export default {
         updateUser(state, payload) {
             state.user = payload;
         },
-        setTemplate(state, payload) {
+        setTemplates(state, payload) {
+            state.templates = payload;
+        },
+        setSelectedTemplate(state, payload) {
             state.selectedTemplate = payload;
         }
     },
@@ -39,6 +43,7 @@ export default {
         async auth(context, { mode, data }) {
             const response = await axios.post("/api", { [mode]: data });
             const { access_token: token, user, exp } = response.data;
+            await context.dispatch("loadTemplates", { user, token });
             context.commit("setUser", { user, token });
 
             localStorage.setItem("user", JSON.stringify(user));
@@ -48,23 +53,26 @@ export default {
             const timeLeft = new Date(exp).getTime() - Date.now();
             logoutTimer = setTimeout(() => context.commit("logout"), timeLeft);
 
-            const { io, user: loggedInUser } = context.getters;
-            io.emit("joinRoom", loggedInUser.id);
+            const { io } = context.getters;
+            io.emit("joinRoom", user.id);
         },
         tryAutoLogin(context) {
             const user = JSON.parse(localStorage.getItem("user"));
             const { token, expirationDate } = localStorage;
-            if (!user || !token || !expirationDate) return;
+            const templates = JSON.parse(localStorage.getItem("templates"));
+
+            if (!user || !token || !expirationDate || !templates) return;
             const timeLeft = new Date(expirationDate).getTime() - Date.now();
             if (timeLeft <= 0) {
                 localStorage.clear();
                 return;
             }
             context.commit("setUser", { user, token });
+            context.commit("setTemplates", templates);
             logoutTimer = setTimeout(() => context.commit("logout"), timeLeft);
 
-            const { io, user: loggedInUser } = context.getters;
-            io.emit("joinRoom", loggedInUser.id);
+            const { io } = context.getters;
+            io.emit("joinRoom", user.id);
         },
         logout(context) {
             context.commit("removeUser");
@@ -81,9 +89,17 @@ export default {
             context.commit("updateUser", user);
             localStorage.setItem("user", JSON.stringify(user));
         },
+        async loadTemplates(context, { user, token }) {
+            const { data: { templates } } = await axios.post(
+                "/xapi",
+                { userID: user.id, getTemplateNames: true },
+                { headers: { Authorization: `Bearer ${token}` } }
+            );
+            context.commit("setTemplates", templates);
+            localStorage.setItem("templates", JSON.stringify(templates));
+        },
         selectTemplate(context, template) {
-            context.commit("setTemplate", template);
-            localStorage.setItem("selectedTemplate", template);
+            context.commit("setSelectedTemplate", template);
         }
     },
     getters: {
@@ -95,6 +111,9 @@ export default {
         },
         user(state) {
             return state.user;
+        },
+        templates(state) {
+            return state.templates;
         },
         selectedTemplate(state) {
             return state.selectedTemplate;
